Amend code as per PR comments
